requirements.txt
torch
torchvision
monai
nibabel
numpy
scipy
streamlit
openai
python-dotenv
```
